requirements.txt
azure-common==1.1.28
azure-core==1.22.1
azure-identity==1.6.0
azure-keyvault-secrets==4.3.0
azure-storage-file-share==12.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
mysql-connector-python==8.0.28
oauthlib==3.2.0
portalocker==2.3.2
protobuf==3.19.4
pycparser==2.21
PyJWT==2.3.0
pyodbc==4.0.32
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
Unidecode==1.3.2
urllib3==1.26.8
